+++ trinity/algorithm/sample_strategy/__init__.py
@@ -1,4 +1,3 @@
-from trinity.algorithm.sample_strategy.mix_sample_strategy import MixSampleStrategy
 from trinity.algorithm.sample_strategy.sample_strategy import (
     SAMPLE_STRATEGY,
     DefaultSampleStrategy,
